+++ src/main/java/nz/co/mirality/colony4cc/client/CustomRenderTypes.java
@@ -1,0 +1,50 @@
+package nz.co.mirality.colony4cc.client;
+
+import com.mojang.blaze3d.platform.GlStateManager;
+import com.mojang.blaze3d.systems.RenderSystem;
+import net.minecraft.client.renderer.RenderState;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
+import net.minecraft.client.renderer.vertex.VertexFormat;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import org.lwjgl.opengl.GL11;
+
+import java.util.OptionalDouble;
+
+@OnlyIn(Dist.CLIENT)
+public class CustomRenderTypes extends RenderType
+{
+    // Dummy
+    private CustomRenderTypes(String name, VertexFormat format, int p_i225992_3_, int p_i225992_4_, boolean p_i225992_5_, boolean p_i225992_6_, Runnable runnablePre, Runnable runnablePost) {
+        super(name, format, p_i225992_3_, p_i225992_4_, p_i225992_5_, p_i225992_6_, runnablePre, runnablePost);
+    }
+
+    private static final LineState THICK_LINES = new LineState(OptionalDouble.of(3.0D));
+
+    public static final RenderType OVERLAY_LINES = create("overlay_lines",
+            DefaultVertexFormats.POSITION_COLOR, GL11.GL_LINES, 256,
+            RenderType.State.builder()
+                    .setLineState(THICK_LINES)
+                    .setLayeringState(VIEW_OFFSET_Z_LAYERING)
+                    .setTransparencyState(TRANSLUCENT_TRANSPARENCY)
+                    .setTextureState(NO_TEXTURE)
+                    .setDepthTestState(NO_DEPTH_TEST)
+                    .setCullState(NO_CULL)
+                    .setLightmapState(NO_LIGHTMAP)
+                    .setWriteMaskState(COLOR_WRITE)
+                    .createCompositeState(false));
+
+    public static final RenderType CLAIM_AREA = create("claimed_area",
+            DefaultVertexFormats.POSITION_COLOR, GL11.GL_QUADS, 256, false, true,
+            RenderType.State.builder()
+                .setTransparencyState(new TransparencyState("translucent_transparency", () -> {
+                    RenderSystem.enableBlend();
+                    RenderSystem.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
+                    RenderSystem.enableAlphaTest();
+                }, () -> {
+                    RenderSystem.disableBlend();
+                    RenderSystem.disableAlphaTest();
+                }))
+                .createCompositeState(true));
+}