+++ src/main/java/nz/co/mirality/colony4cc/peripheral/TileColonyPeripheral.java
@@ -1,8 +1,12 @@
 package nz.co.mirality.colony4cc.peripheral;
 
 import dan200.computercraft.api.peripheral.IPeripheral;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.Direction;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.minecraftforge.items.CapabilityItemHandler;
+import net.minecraftforge.items.IItemHandler;
 import nz.co.mirality.colony4cc.block.PeripheralTile;
 
 import javax.annotation.Nullable;
@@ -37,5 +41,17 @@ public class TileColonyPeripheral extends ColonyPeripheral {
     @Override
     public BlockPos getPos() {
         return this.tile.getBlockPos();
+    }
+
+    @Override
+    @Nullable
+    protected IItemHandler getInventory(@Nullable Direction side)
+    {
+        if (side == null) side = Direction.UP;
+        final BlockPos pos = getPos().relative(side);
+        final TileEntity target = getWorld().getBlockEntity(pos);
+        if (target == null) return null;
+
+        return target.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY, side.getOpposite()).orElse(null);
     }
 }